backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.2
scikit-learn>=1.4.2
pandas>=2.0.0
numpy>=2.0.0
joblib>=1.3.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
groq>=0.4.1
python-jose[cryptography]
passlib[bcrypt]
bcrypt
slowapi>=0.1.9
pytest>=7.4.0
imbalanced-learn>=0.11.0
matplotlib>=3.8.0
seaborn>=0.13.0

